Moved zygote utils test to separate file
